Add code to trim word2vec model down to the game vocabulary
The game vocabulary has 40K words, which is not that much.

The pre-trained GoogleNews word2vec model has 3M words, which is a lot, in fact too much to be loaded using 64 bit python. (And I couldn't get the nltk package that the paltform requires to run on 64 bit).

Therefore, out of performance+necessity, we want to trim down any word embeding we use to the game vocabulary.