First working version of GloomyBot!
Super simple strategy: try to get a clue for as many of your words as possible, with the condition that the "weakest" of your words must maintain a "safety margin" to the undersired words (the margin is variable so that we can avoid the assasin more than neutral words). Thresholds are very very coarsely tuned.

It's not amazing. I've only looked at a small sample of clues it proposes so too early to drawn conclusions, but my hunch is that the majority of the gains are to be had from improving the semantic qualities of the embedding, not from playing a fancier strategy.